We should never have been using the `jacobian` function to estimate jacobian-vector products. Instead, one should estimate the jvp directly by estimating the derivative in the direction of `ẋ`. It's definitely faster, and probably more numerically stable.